requirements.txt
kafka-python
cryptography